refactor: existing issues modal code
